project_root_directory/requirements.txt
Django == 5.1.6
djangorestframework == 3.15.2
python-dotenv == 1.0.1
pillow==11.1.0
django-filter==25.1
PyJWT==2.10.1
mysqlclient==2.2.7
twilio==6.38.0
africastalking==1.2.9